requirements/prod.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.1
python-decouple==3.3
psycopg2-binary==2.8.6
honcho==1.0.1
gunicorn==19.9.0
django-cors-headers==2.4.0
Markdown==3.3.1
django-filter==2.4.0
djangorestframework==3.12.1
PyYAML==5.3.1
coreapi==2.3.3
django-rest-auth==0.9.5